qmemman: separate data struct from implementation into modules
This helps avoiding cyclic imports - some places (especially
qubes.vm.qubesvm) needs only some of the parts).
